Ticket Discussions refactor
